add a common AzureKeyVaultKeyStoreUtil class
